backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
langchain
langchain-community
langchain-google-genai
google-generativeai